Shortrocity - childstories/requirements.txt
openai
python-dotenv
deepgram-sdk
together
opencv-python
pydub
# ffmpeg-python

g4f[all]
elevenlabs